Rename some YXML_ tokens for consistency + add token for end-of-attr
Previously, reading the value of an attribute required reading all DATA
tokens until the next YXML_ATTR or the next YXML_EOA. The new
YXML_ATTRSTART and YXML_ATTREND are more clear and provides a more
tighter bound as to where you can stop waiting for more DATA tokens for
the attribute value.

The YXML_EOA token has been renamed to YXML_CONTENT, because that's what
it actually signifies: The start of the contents of the element.

I've also documented the token ranges in which the elem, data and attr
fields in yxml_t remain valid.