Add 'misc3' state to handle 'Misc' data after the root element closed
Previously the misc2 state was entered after the root element has
closed, which would still allow for character content and new open tags
to be parsed. The latter was already detected with the 'afterelem' trick
added in 6bc21882f (now removed again), but that commit did not disallow
character content. This also removes the YXML_EMULROOT error code, such
errors are now reported as YXML_ESYN.

This commit adds ~600 bytes and improves performance for one benchmark
and worsens performance for the other. Neither difference is very
significant, however.