Allow non-ASCII characters in attribute and element names
Similar reasoning as for allowing non-ASCII characters in data: We can't
validate them anyway because yxml operates on bytes and is unaware of
the encoding. This does allow a wide range of characters as
element/attribute names that aren't formally allowed, but the most
common use of those names in applications is simply to check whether a
particular element/attribute name matches one that it knows, and unknown
names are generally ignored.

Without this change, it is impossible to parse "international" XML
documents with yxml. It is possible now, but applications do need to do
further validation on their own if they want to be conforming.